removing states from RotorUtils, and creating RotorCtl to house states
