--- app/src/test/java/ai/rotor/rotorvehicle/BluetoothServiceTest.java
@@ -1,23 +1,0 @@
-package ai.rotor.rotorvehicle;
-
-import android.content.Context;
-
-import org.junit.Before;
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-public class BluetoothServiceTest {
-    private BluetoothService bluetoothService;
-
-    @Before
-    public void setup() {
-        bluetoothService = new BluetoothService();
-    }
-
-    @Test
-    public void setCorrectInitialState() {
-
-    }
-
-}
